+++ Private/Test-Volume.ps1
@@ -3,17 +3,19 @@ function Test-Volume {
     param (
         [Parameter(Mandatory = $true)]
         [ValidatePattern('^[A-Za-z]$')]
-        [String]$DriveLetter
+        [String]$DriveLetter,
+
+        [Switch]$Force
     )
     $DriveLetter = $DriveLetter.Trim().Substring(0, 1).ToUpperInvariant()
     $GetVolume = Get-Volume -DriveLetter $DriveLetter -ErrorAction SilentlyContinue
     if (-not $GetVolume) {
         return $false
     }
-    if (Test-SystemVolume -DriveLetter $DriveLetter) {
+    if (-not $Force -and (Test-SystemVolume -DriveLetter $DriveLetter)) {
         return $false
     }
-    if (Test-BootVolume -DriveLetter $DriveLetter) {
+    if (-not $Force -and (Test-BootVolume -DriveLetter $DriveLetter)) {
         return $false
     }
     if (-not (Test-PartitionType -DriveLetter $DriveLetter)) {